+++ CHANGELOG.md
@@ -18,11 +18,13 @@ CHANGED:
 - major version to 3.*
 - update of gradle-plugin dependency with id 'io.freefair.lombok' to new patch version 8.10
 - update of gradle-plugin dependency with id 'com.diffplug.spotless' to new beta version 7.0.0.BETA2
+- update of dependency file-worker to new version 18.0
 - update of dependency silly-bean to new major version 3.0
 - update of dependency silly-collection to new major version 28.1
+- update of dependency silly-strings to new version 9.2
 - update of test dependency commons-lang3 to new version 3.17.0
 - update of test dependency test-object to new major version 9
-- update of test dependency mockito-core to new version 5.14.1
+- update of test dependency mockito-core to the new version 5.14.1
 - update of test dependency junit-jupiter-api to new version 5.11.2
 - dependencies are managed now over bundles